+++ src/components/all.sass
@@ -306,8 +306,28 @@ $offer-card-padding: 3rem
   max-width: 360px
   box-shadow: 0px 12px 30px 0 rgba(22, 22, 22, 0.06)
   background-color: #ffffff
-  padding: $offer-card-padding 0
+  padding: 0 0 $offer-card-padding
   text-align: center
+
+  &__mainInfoContainer
+    padding-top: $offer-card-padding
+    padding-bottom: 1.5rem
+    position: relative
+    overflow: hidden
+
+  &__mainInfoBackground
+    display: block
+    position: absolute
+    z-index: 0
+    top: 0
+    bottom: 0
+    width: 130%
+    left: -15%
+    border-radius: 0 0 50% 50%
+
+  &__mainInfo
+    position: relative
+    z-index: 1
 
   ul
     display: inline-block
