+++ src/content/plans.md
@@ -2,12 +2,14 @@
 content-key: plans
 
 plans:
-  - name: Basic
-    amount: 1000
+  - name: Early Adopter
+    amount: 350000
+    terms: Per user, per year
     features: ['60 day free trial', 'All functionality', 'Dedicated subdomain', 'Up to 20 repositories', 'Email and phone support']
     color: "#fba92c"
   - name: Professional
-    amount: 1500
+    amount: 450000
+    terms: Per user, per year
     features: ['60 day free trial', 'All functionality', 'Dedicated subdomain', 'Up to 20 repositories', 'Email and phone support']
     color: "#58abff"
 ---
